refactor: remove redundant reducer files and import directly from Reviews.reducer
- Remove Reviews/Reviews.reducer.ts, Overrated/Overrated.reducer.ts,
  Underrated/Underrated.reducer.ts, and Underseen/Underseen.reducer.ts
- Update components to import directly from Reviews.reducer.ts
- Eliminates unnecessary indirection and simplifies the codebase
