--- src/components/Reviews/Reviews.reducer.ts
@@ -1,6 +1,0 @@
-export { 
-  Actions, 
-  initState,
-  reducer,
-  type Sort
-} from "~/components/Reviews.reducer";
